Replace double with BigDecimal

+++ src/main/java/com/litesoftwares/coingecko/domain/Coins/CoinData/CommunityData.java
@@ -2,6 +2,7 @@ package com.litesoftwares.coingecko.domain.Coins.CoinData;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import java.math.BigDecimal;
 import lombok.*;
 
 @Data
@@ -12,13 +13,13 @@ public class CommunityData {
     @JsonProperty("twitter_followers")
     private long twitterFollowers;
     @JsonProperty("reddit_average_posts_48h")
-    private double redditAveragePosts48h;
+    private BigDecimal redditAveragePosts48h;
     @JsonProperty("reddit_average_comments_48h")
-    private double redditAverageComments48h;
+    private BigDecimal redditAverageComments48h;
     @JsonProperty("reddit_subscribers")
     private long redditSubscribers;
     @JsonProperty("reddit_accounts_active_48h")
-    private double redditAccountsActive48h;
+    private BigDecimal redditAccountsActive48h;
     @JsonProperty("telegram_channel_user_count")
     private long telegramChannelUserCount;
